fix: anchor engine gui-free CI grep to CTest + shrink channel enum tags
CI's "Enforce engine GUI-free" step used an unanchored grep that matched
`#include <juce_gui_basics/...>` inside EngineGuiGuard.h's explanatory
comments (lines 15/28/42), failing the build job on a false positive.
Anchor the regex to line start so it mirrors tests/CheckEngineGuiFree.cmake
exactly — real include directives are still caught, comment mentions ignored.

Also shrink EngineCommandType / EngineEventType from uint16_t to uint8_t to
clear clang-tidy performance-enum-size (both fit easily; FIFO payloads are
size-agnostic, static_asserts and layout unaffected, nothing serialized).

Test Plan
- Anchored grep clean on engine/ (0 matches); still matches a real GUI
  include (verified with a temp file, removed).
- 26/26 default ctest; 20/20 TSan (-L unit); clang-tidy performance-enum-size
  resolved on both headers.

+++ engine/graph/EngineEvent.h
@@ -12,7 +12,7 @@ namespace arpbox::engine
     event half). Where `EngineSnapshot` carries continuous per-block STATE, these
     are one-shot NOTIFICATIONS the UI reacts to once (fire an animation, show a
     banner, recompute a latency badge). */
-enum class EngineEventType : std::uint16_t
+enum class EngineEventType : std::uint8_t
 {
     deviceChanged = 0,       ///< Active audio device changed (a = new SR, b = block size).
     deviceDied,              ///< Current device dropped out (unplugged / lost).
